FSPT-561: move managed expression builder config out of helper
